requirements.txt
httpx
python-dotenv
pytest
pytest-mock